pyproject.toml
[tool.poetry]
name = "nba-analysis"
version = "0.1.0"
description = ""
authors = ["jacasta2 <jacasta2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pip = "^24.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"
pandas = "2.0.3"
numpy = "^1.26.4"
nptyping = "^2.5.0"
nba-api = "^1.4.1"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
scikit-learn = "^1.4.1.post1"
statsmodels = "^0.14.1"
shap = "^0.44.1"
dataframe-image = "^0.2.3"
hopsworks = "3.4.4"
ipywidgets = "^8.1.2"
streamlit = "^1.31.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
